Add task type.
Consists of title and status.
Status is implemented as discriminated union. The alternative is a list of status strings. The advantage of discriminated union is simplicity and a good, built-in type. The advantage of list of status strings is a built-in string to use in user interfaces, and ability to add and remove statuses. Might consider changing to it later.

Change from crlf to lf line ending.